Generate the set of unique combinations of tide ID, motus tag, and receiver.
Valid tide IDs are determined per tag, based on the first and last detection time of the tag.

Join the complete dataset (all combinations) with the detection data, and fill the rest (periods with no detections) with zero values.

IMPORTANT NOTE / TO DO:
Receiver uptime has not yet been checked implemented. (i.e., the data will include tidal periods for receivers even when the receiver was not working)