Improved plot.addSunriseSet function.
Previously, the function required information about the dates for the plot to be in the global environment.

Now that I have realised you can access the underlying data for a plot (p$data), the only required input is the plot object p.

Also, it actually works properly now - finding sunset and sunrise for ALL dates of the plot, not just those present in df.alltags.

ALSO: Finds sunrise/set based on specific Newcastle coordinates, so that it is consistent across stations for simplicity.